react font awesome added

+++ src/Pages/Dashboard/MyOrders/MyOrders.js
@@ -1,6 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import SingleOrder from '../SingleOrder/SingleOrder';
 import useAuth from '../../../hooks/useAuth'
+
+
 const MyOrders = () => {
     const [orders, setOrders] = useState([])
     const { user } = useAuth()
@@ -15,18 +17,22 @@ const MyOrders = () => {
     }, [user])
 
     const handleCancel = (id) => {
-        console.log(id)
-        fetch(`https://glacial-bastion-50505.herokuapp.com/deleteProduct/${id}`, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount) {
-                    alert('Order canceled successfully')
-                    const newOrder = orders.filter(order => order._id !== id)
-                    setOrders(newOrder)
-                }
+
+        const success = window.confirm('Are you sure you want to cancel this order ?')
+        if (success) {
+            fetch(`https://glacial-bastion-50505.herokuapp.com/deleteProduct/${id}`, {
+                method: 'DELETE'
             })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount) {
+                        alert('Order canceled successfully')
+                        const newOrder = orders.filter(order => order._id !== id)
+                        setOrders(newOrder)
+                    }
+                })
+        }
+
 
     }
     return (
